Find2.0 changes for the Quant  and De-Convolution (#1408)
* use find2.0 for the convolution

(cherry picked from commit 5fa42993b7670b4a9c48cf9d5a73f88a630b2d78)
